% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.sdem}
\alias{plot.sdem}
\title{Basic summary of objects of class 'sdem'}
\usage{
\method{plot}{sdem}(object, extended = FALSE, ...)
}
\arguments{
\item{extended}{logical. if TRUE additional information is printed}
}
\value{
A huge amount of information
}
\description{
Basic summary of objects of class 'sdem'
}
