% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R, R/utils.R
\name{summary.sdem}
\alias{summary.sdem}
\title{Summary of sdem class}
\usage{
\method{summary}{sdem}(object, correlation = FALSE)

\method{summary}{sdem}(object, correlation = FALSE)
}
\arguments{
\item{object}{an object of class 'sdem'.}

\item{extended}{logical. if 'TRUE' the parameter correlation matrix is also printed.}

\item{...}{further arguments passed to/from other methods.}
}
\value{
a lot of information from the optimisation. parameter estimates,
state estimates etc.

A huge amount of information
}
\description{
'summary.sdem' prints parameter estimates, standard deviations,
the slope of the objective function in parameter directions etc.
}
